requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10.1
Flask-Cache==0.13.1
python-dateutil==2.2
LinkHeader==0.4.2
requests==2.5.0
jira==0.16
feedparser==5.1.3
more-itertools==2.2
